pyproject.toml
[tool.poetry]
name = "credentials"
version = "0.1.0"
description = ""
authors = ["mr-mapache"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.0"
psycopg = "^3.2.3"
uvicorn = "^0.34.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
fastapi = {extras = ["standard"], version = "^0.115.11"}
bcrypt = "^4.3.0"
itsdangerous = "^2.2.0"


[tool.poetry.group.tests.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
